backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.27.0
itsdangerous==2.2.0
python-dotenv==1.0.1
cryptography==41.0.7
